Corrigiendo errores de calendario y agregando login modal
