+++ client/components/ReservaModal.tsx
@@ -5,15 +5,13 @@ import { InlineDatePickerWithRange } from "./CalendarModal";
 import { useRouter } from "next/navigation";
 
 
-const BookingModal = ({ isOpen, onClose, orders = [], isAuthenticated }) => {
+const BookingModal = ({ isOpen, onClose, orders = [], product }) => {
   const [selectedDates, setSelectedDates] = useState({
     from: null,
     to: null,
   });
 
   const router = useRouter();
-
-  const [showLoginPrompt, setShowLoginPrompt] = useState(!isAuthenticated);
 
   // Procesar las órdenes como rangos deshabilitados
   const disabledDates = orders.map((order) => ({
@@ -34,13 +32,10 @@ const BookingModal = ({ isOpen, onClose, orders = [], isAuthenticated }) => {
       alert("Las fechas seleccionadas están reservadas.");
     }
   }
-  
-  const handleNavigation = (path: string) => {
-    router.push(`${path}`);
-  };
 
   if (!isOpen) return null;
-
+ console.log(product);
+ 
 
   return (
     <div className="container1 fixed inset-0 flex items-end sm:items-center sm:justify-center bottom-0 bg-black bg-opacity-50 z-50">
@@ -54,45 +49,26 @@ const BookingModal = ({ isOpen, onClose, orders = [], isAuthenticated }) => {
             &times;
           </button>
         </div>
-        {showLoginPrompt ? (
-          <div className="flex flex-col place-content-center gap-2  h-60 text-center ">
-            <h3 className="text-lg font-semibold mb-4">
-              Identifícate para reservar
-            </h3>
-            <p className="text-sm text-gray-600 mb-4">
-              Ingresa a tu cuenta y reserva todos los artículos que necesites
-              para tu fiesta.
-            </p>
-            <button className="bg-primary text-white px-4 py-2 rounded w-full mb-2" onClick={() => handleNavigation("/login")}>
-              Iniciar sesión
-            </button>
-            <button className="bg-secondary text-white px-4 py-2 rounded w-full" onClick={() => handleNavigation("/signup")}>
-              Crear cuenta
-            </button>
+        <div className="container1.1.2 flex w-full gap-2 sm:gap-3">
+          <div className="container1.1.2.1 mx-auto relative flex flex-col items-center w-full justify-center">
+            <InlineDatePickerWithRange
+              date={selectedDates}
+              onDateChange={handleDateChange}
+              disabledDates={disabledDates}
+              product={product}
+            />
           </div>
-        ) : (
-          <>
-            <div className="container1.1.2 flex w-full gap-2 sm:gap-3">
-              <div className="container1.1.2.1 mx-auto relative flex flex-col items-center w-full justify-center">
-                <InlineDatePickerWithRange
-                  date={selectedDates}
-                  onDateChange={handleDateChange}
-                  disabledDates={disabledDates}
-                />
-              </div>
-            </div>
-            <button
-              className={`w-full py-2 rounded-md ${selectedDates.from && selectedDates.to
-                ? "bg-secondary text-white"
-                : "bg-gray-400 text-white cursor-not-allowed"
-                }`}
-              disabled={!selectedDates.from || !selectedDates.to}
+        </div>
+        <button
+          className={`w-full py-2 rounded-md ${selectedDates.from && selectedDates.to
+            ? "bg-secondary text-white"
+            : "bg-gray-400 text-white cursor-not-allowed"
+            }`}
+          disabled={!selectedDates.from || !selectedDates.to}
 
-            >
-              Continuar reserva
-            </button>
-          </>
-        )}
+        >
+          Continuar reserva
+        </button>
       </div>
     </div>
   );
